pyproject.toml
[project]
name = "multi-doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "einops>=0.8.2",
    "faiss-cpu>=1.13.2",
    "groq>=1.0.0",
    "pathlib>=1.0.1",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "transformers>=5.0.0",
]
